updated model classification to be pre and post radiative transfer calculation
