+++ archnemesis/Models/PostRTModels.py
@@ -1,5 +1,9 @@
 from __future__ import annotations #  for 3.9 compatability
 
+"""
+Contains models that alter the replica after radiative transfer has been calculated
+"""
+
 from .ModelBase import *
 
 
@@ -9,7 +13,7 @@ _lgr.setLevel(logging.DEBUG)
 
 
 
-class SpectralModelBase(ModelBase):
+class PostRTModelBase(ModelBase):
     """
     Abstract base class of all parameterised models used by ArchNemesis that interact with the calculated spectrum in the forward model.
     """
@@ -65,7 +69,7 @@ class SpectralModelBase(ModelBase):
         raise NotImplementedError(f'calculate_from_subspecret should be implemented for all Spectral models')
 
 
-class TemplateSpectralModel(SpectralModelBase):
+class TemplateSpectralModel(PostRTModelBase):
     """
         This docstring acts as the description for the model, REPLACE THIS.
     """
@@ -366,7 +370,7 @@ class TemplateSpectralModel(SpectralModelBase):
         return
 
 
-class Model231(SpectralModelBase):
+class Model231(PostRTModelBase):
     """
         Scaling of spectrum using a varying scaling factor (following a polynomial of degree N)
         
@@ -519,7 +523,7 @@ class Model231(SpectralModelBase):
 
 
 # [JD] This uses forward_model.SpectroscopyX, whereas Model231 uses forward_model.Measurement, is this correct?
-class Model2310(SpectralModelBase):
+class Model2310(PostRTModelBase):
     """
         Scaling of spectra using a varying scaling factor (following a polynomial of degree N)
         in multiple spectral windows
@@ -676,7 +680,7 @@ class Model2310(SpectralModelBase):
         )
 
 
-class Model232(SpectralModelBase):
+class Model232(PostRTModelBase):
     """
         Continuum addition to transmission spectra using the angstrom coefficient
 
@@ -812,7 +816,7 @@ class Model232(SpectralModelBase):
             _lgr.warning(f'It looks like there is no calculation for NGEOM=1 for model id = {self.id}')
 
 
-class Model233(SpectralModelBase):
+class Model233(PostRTModelBase):
     """
         Continuum addition to transmission spectra using a variable angstrom coefficient (Schuster et al., 2006 JGR)
 
@@ -981,7 +985,7 @@ class Model233(SpectralModelBase):
             )
 
 
-class Model667(SpectralModelBase):
+class Model667(PostRTModelBase):
     """
         In this model, the output spectrum is scaled using a dillusion factor to account
         for strong temperature gradients in exoplanets
