+++ archnemesis/Models/__init__.py
@@ -2,9 +2,8 @@ from __future__ import annotations #  for 3.9 compatability
 
 from typing import Iterator
 
-from . import AtmosphericModels
-from . import NonAtmosphericModels
-from . import SpectralModels
+from . import PreRTModels
+from . import PostRTModels
 from .ModelParameterEntry import ModelParameterEntry
 from .ModelBase import ModelBase
 
@@ -23,19 +22,13 @@ def _get_all_model_classes():
     return (
         *tuple(
             x[1] for x in inspect.getmembers(
-                AtmosphericModels,
+                PreRTModels,
                 lambda x: inspect.isclass(x) and issubclass(x, ModelBase) and x.id is not None
             )
         ),
         *tuple(
             x[1] for x in inspect.getmembers(
-                NonAtmosphericModels,
-                lambda x: inspect.isclass(x) and issubclass(x, ModelBase) and x.id is not None
-            )
-        ),
-        *tuple(
-            x[1] for x in inspect.getmembers(
-                SpectralModels,
+                PostRTModels,
                 lambda x: inspect.isclass(x) and issubclass(x, ModelBase) and x.id is not None
             )
         ),
